move preferences button
