+++ src/components/Footers.tsx
@@ -1,7 +1,6 @@
 import {
   Button,
   Flex,
-  Image,
   Link,
   Modal,
   ModalBody,
@@ -9,88 +8,11 @@ import {
   ModalFooter,
   ModalHeader,
   ModalOverlay,
-  Select,
-  Spinner,
   Text,
-  Tooltip,
 } from "@chakra-ui/react";
-import { useRef, useState } from "react";
+import { useState } from "react";
 
-import { COLOR_SCHEME_PRESETS } from "../lib/colors";
 import { State } from "../lib/state";
-import { DEFAULT_PREFERENCES, Preferences } from "../lib/schema";
-
-function PreferencesModal(props: {
-  state: State;
-  preferences: Preferences;
-}) {
-  const { preferences: originalPreferences, state } = props;
-  const [visible, setVisible] = useState(false);
-  const [preferences, setPreferences] = useState(DEFAULT_PREFERENCES);
-  const initialPreferencesRef = useRef(DEFAULT_PREFERENCES);
-  const initialPreferences = initialPreferencesRef.current;
-
-  const onOpen = () => {
-    initialPreferencesRef.current = originalPreferences;
-    setPreferences(originalPreferences);
-    setVisible(true);
-  };
-
-  const previewPreferences = (newPreferences: Preferences) => {
-    setPreferences(newPreferences);
-    state.setPreferences(newPreferences, false);
-  };
-
-  const onCancel = () => {
-    setPreferences(initialPreferences);
-    state.setPreferences(initialPreferences);
-    setVisible(false);
-  };
-
-  const onConfirm = () => {
-    state.setPreferences(preferences);
-    setVisible(false);
-  };
-
-  return (
-    <>
-      <Button onClick={onOpen}>Preferences</Button>
-      <Modal isOpen={visible} onClose={onCancel}>
-        <ModalOverlay />
-        <ModalContent>
-          <ModalHeader>Preferences</ModalHeader>
-          <ModalBody>
-            <Flex gap={4}>
-              Color scheme:
-              <Select
-                value={preferences.colorScheme.name}
-                onChange={(e) => {
-                  const colorScheme = COLOR_SCHEME_PRESETS.find(
-                    ({ name }) => name === e.target.value
-                  );
-                  if (!colorScheme) return;
-                  previewPreferences({ ...preferences, colorScheme });
-                }}
-              >
-                {COLOR_SCHEME_PRESETS.map(({ name }) => (
-                  <option key={name} value={name}>
-                    {name}
-                  </option>
-                ))}
-              </Select>
-            </Flex>
-          </ModalBody>
-          <ModalFooter>
-            <Button onClick={onCancel} mr={2}>
-              Cancel
-            </Button>
-            <Button onClick={onConfirm}>Save</Button>
-          </ModalFooter>
-        </ModalContent>
-      </Modal>
-    </>
-  );
-}
 
 function AboutModal() {
   const [visible, setVisible] = useState(false);
@@ -137,10 +59,9 @@ function AboutModal() {
 
 /** The footer on the bottom of the calendar. */
 export function LeftFooter(props: {
-  preferences: Preferences;
   state: State;
 }) {
-  const { preferences, state } = props;
+  const { state } = props;
 
   return (
     <Flex
@@ -148,7 +69,6 @@ export function LeftFooter(props: {
       align="center"
       gap={2}
     >
-      <PreferencesModal preferences={preferences} state={state} />
       <Text>Last updated: {state.lastUpdated}.</Text>
       <Flex gap={4}>
         <AboutModal />
